Now pivoting & changing design. PMs will be group
Right now having trouble with writting createMessage and deleteMessage
in the messageController. The problems are code repetition, where to
make checks, and the possibility of there being messages with just
'recepients' and no 'group' and vice versa, is making the logic really
complicated.
I was trying to solve some of these problems by writting mongoose
Schema middleware, but even then it is still complcated/noodle

TODO
Private conversations will now instead ALWAYS be treated as a group of
2 people. This way I dont have to deal with '!!recepients' vs '!!group'.
This will simplify my logic