Almost done barebones
Will now change 'GET /messages/:messageId' and 'POST /messages' to
instead be prefaced by 'chats/:chatId/'.
Why? Well for retriving a message by id, it is not neccessary since
we can retrive the chat id from the resulting message instead. However,
for the POST it is necessary, and we are currently passing in the
chatId via req.body

More importantly though, when creating a route for retrieving all of
the messages that correspond to a sepcific chat, I needed to use the
'GET /chats/:chatId/messages/:messageId' because I do need the chatId.

And in this scenario if I wanted to include the chatId, I would have to
switch to using POST so I can pass req.body, or like I did, include
them in the URL.
But if I leave this the way it is, then I'd have:
'GET /chats/:chatId/messages' and 'GET /messages/:messageId'

This looks very confusing from an API perspective, with some routes
needing the chatId in the body, and others in the param.
I'm not 100% sure, but I think its best to just have all the messages
be prefaced by /chats/:chatId/ like I mentioned