Now having problems with Group Schema/PMs
https://www.mongodb.com/community/
forums/t/scalable-group-chat-modelling-and-performance/124623/2

Should I make changes to my chat/conversation
schema? I'm considering cahngin Chat's members array of userIds, to
an array of [{userId, total_unread}]. This way anytime a new message is
sent, I will increment the total_unread of every user in the group.
Then from the frontend, when a user sees the newest message, it will
drop 'total_unread' back to 0. Also im consider duplicating some data.
In particualar, maybe adding a 'chat/group' array on every user, which
will just be an array that lists which groups this user belongs to.
I'm not sure if this is good though because then it makes it so anytime
a user leaves or is kicked from a group, I have to makes sure to change
both user.groups as well as group.members.
Im guessing the benefit of this is that I can easily query the list of
a user's groups, insetad of running a query on the entire groups
colllection.
I think I will try first, not adding this duplicated(?) field

Also should I separate Group messages vs Private messages by creating
a separate schema, or by creating a separate route?
My initial gut feeling is to create 2 spearate schemas, because I wanted
to apply some restrictions to my mongoDB document. But now I think this
is maybe unnecessary. I only have to worry about the routes that I
expose on  my API, so I can simply accomplish this by having 1 route
to send a private message to a user (or create new one if no conv exist)
then 1 route to create a group, and another route to send a message to
an existing group