Couple fixes. In the middle of chanigng to async
A couple of my controller functions use callbacks, which if they
throw errors, are not properly caught and propelled up to the
asyncHandler, and are then not caught at all by my catch-all handler.

For example using bycript(1param, 2param, () => throw Error)
is not properly caught
Same with jwt.sign i think
Im now in the middle of fixing this

TODO
Change some of the callback code to use promises so that the errors can
be caught by asyncHandler and forwarded to my catch all error handler.
Test all this new things that were added
add prettier and eslint and eslint-plugin-node (?), and typescript,
and add helmet library for security