Improved Error handling format
Changed the format, then added a lot of Error classes.
Basically each error will alwys come with statusCode, message, type
type helps to differentiate errors if I ever need to do this, for ex
on '/sign-up', I can get validation errors on the fields, or I can
get an error when I try to .save() the document in our Database with
Mongoose. Note though that .save() also runs mongoose's validation
methods if we have any

TODO
-Will now improve the way memeberships work. Currently, I am loading all
the available memebrships on startup, and then keeping a reference to
this array. I use this to easily access them when a new user signs up.
However, I now think this is a terrible idea, because if I update
a membership in the database, then this array that was loaded on
startup WONT be update, which the whole app and every new user will get
outdated memberships objects. The id will still be correct if I haven't
changed it, but the rest of it will be innacurate.
So I think its probably best to just no longer use a static array on
startup, and instead fetch for the memeberships everytime a new user
is create.
-Test the rest
-add prettier and eslint and eslint-plugin-node (?), and typescript,
-and add helmet library for security