Updated Memberships, Extracted chat & validation
-Memberships now update on change stream. So whenever a change happens
to the Memberships collection, a handler updates our cached array of
memebrships. We have a getter function that retrieves a copy of this
cached memebrships array, which we use to issue memberships when a new
user is created, but will 99% also use when displaying a list of
memebrships to the user.
This way instead of always making a database request for all
memberships when creating a new user or when displaying memebrships,
I can instead just always show this cache which updates anytime
the Memebrships collection changes (once in a blue moon)
-Extracted chat & Validation middlewares
-Fixed some errors in controllers & Added TransactionError
-Added a few missing checks in message and chat Controllers

TODO
-Test the rest (message and chat controllers and routes)
-add prettier and eslint and eslint-plugin-node (?), and typescript,
-and add helmet library for security