Writting createChat - next change Chat and Members
-Fixed CustomError's undefined props scenario by adding ' = {}' to its
object props
- wrote getChats , replacing getChat. getChats is meant for when
a user wants to load their opened conversations

TODO
-Will now change chat Schema and Memberships. I finished writting a
first draft of createChat. I realized this needs to handle different
types of chat creations: global, group, private. I noticed the
'isGroupChat' body parameter is not needed, because this can be derived
from the amount of valid members passed in. If there are 2 total valid
members (including self), then its private message, otherwise its group
chat. In a similar idea, we can consolidate 'isGroupChat' and 'isGlobal'
from the ChatSchema, into just one field of type enum String, that
indicates if type is global, private, or group.
I also think its best to change how we handle our memberships. Currently
we are holding and handling memberships in an inner array inside of
Chat, and we also have an array of chats inside of User.
Not sure what I'll do with the array inside of user, but for ChatSchema,
I think its just best to extract memebrships into its own collection.
For ex: ChamMembershipSchema({chat, user}). This collection has a new
document for each and every membership accross any memberships.
Why? Because I think I do want to handle Global Chats, but I dont want
to impose a memebers limit on it (which I would be forced to do because
of the mondoDB limit on doucment size caused by unbounded arrays).
And I think for these global chats, its best to just have one unified
'ChatSchema' that can hanbdle either of the 3 types of chats, instead of
one ChatSchema with the embedded members array and a members limit, and
one GlobalChatSchema, with the separate ChamMembershipSchema.
This unified chat solution feels a lot more future-proof
-Test the rest (message and chat controllers and routes)
-add prettier and eslint and eslint-plugin-node (?), and typescript,
-and add helmet library for security