Decided to scrap all embeded unbounded arrays
Changed the Architecture to use UserChat and UserMessage collections.
Every document here stores the information that corresponds to these
combination of 2 things.

Decided to use this instead of embedding this information inside of
unbounded arrays, because I foresaw the need to add muliple things to
the 'UserChat' (whether embedded inside Chat or as its own collection).
For example, maybe a role or isAdmin or something, and maybe 1 other
field. And with the embedded approach, the problem is the more fields
and data I add, the less members that I can support.
I was also concerned about running into issues with race conditions or
other problems with multiple users constantly editting the same 1 chat
document, for example when they indicate their last read message from
the chat to later on calculate how many new messages there are. I am
concerned that many users rewritting this one document in a short
amount of time can run into trouble

I had some other concerns as well, but same thing with UserMessage.
So I decided to just turjn them into their own collections, with a new
document for each UserChat or UserMessage combination

I rewrote all the code in chatController as a first draft. Need to test
this. Have not yet rewritten messageController, need to do this.
Theres a lot of comments to delete, will do this in the next commit

TODO
-rewrite messageController with the new Architecture
-delete a lot of the comments/old code
- Delete users.chats, I kept this by accident
-Test the rest
-add prettier and eslint and eslint-plugin-node (?), and typescript,
-and add helmet library for security